Added rabbitmq configurations

+++ src/main/java/com/mindstix/microservices/foundation/transactionservice/model/TransactionData.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -16,5 +17,5 @@ public class TransactionData {
     @NotNull
     String transactionType;
     @NotNull
-    Double amount;
+    BigDecimal amount;
 }
